feat(kernel): Add object model with Meta and Trait
`@floriver/kernel`'s object model is a framework for creating extensible types using pure functions, while maintaining natural syntax, without relying on JavaScript's quasi-OOP features. It is by nature an opinionated way of writing types and is meant to be used by libraries and application models rather than for web APIs and DTOs. It is inspired by Rust's `impl` blocks and uses a model similar to Rust's traits for implementing common functionality and extending the behavior of existing types.